DigiFerro/requirements.txt
opencv-python
pyqt
numpy
pandas
